functions/requirements.txt
azure-functions
openai>=1.12.0
azure-search-documents>=11.4.0
azure-cosmos>=4.5.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
requests>=2.31.0
